+++ internal/features/educators_periods/events/command_remove.go
@@ -76,7 +76,7 @@ func loadRemoveEducatorFromPeriodContext(
 	*removeEducatorFromPeriodContext,
 	error,
 ) {
-	query := streamQuery(periodID, educatorID)
+	query := educatorPeriodStreamQuery(periodID, educatorID)
 	events, err := retriever.GetEvents(ctx, eventstore.NoEventPosition, 100, eventstore.Forward, query)
 	if err != nil {
 		return nil, err