+++ internal/features/periods/dto/form_view.go
@@ -4,6 +4,7 @@ import (
 	"seek/internal/features/_shared/shareddto"
 	"seek/internal/features/_shared/sharedmodels"
 	edto "seek/internal/features/educators/dto"
+	emodels "seek/internal/features/educators/models"
 	"seek/internal/features/periods/events"
 	"seek/internal/features/periods/models"
 	sdto "seek/internal/features/students/dto"
@@ -12,41 +13,42 @@ import (
 )
 
 type PeriodFormView struct {
-	ID          string                   `json:"id"`
-	Title       string                   `json:"title"`
-	ServiceType sharedmodels.ServiceType `json:"service_type"`
-	StartTime   sharedmodels.TimeOnly    `json:"start_time"`
-	EndTime     sharedmodels.TimeOnly    `json:"end_time"`
-	Duration    int                      `json:"duration"`
-	Days        shareddto.DaysFormView   `json:"days"`
-	EducatorIDs string                   `json:"educator_ids"`
-	StudentIDs  string                   `json:"student_ids"`
-	URL         string                   `json:"url"`
-	Validation  map[string]events.Validation
-	StudentList []sdto.StudentSelectBoxView
-	Educators   []edto.EducatorView
+	ID              string                   `json:"id"`
+	Title           string                   `json:"title"`
+	ServiceType     sharedmodels.ServiceType `json:"service_type"`
+	StartTime       sharedmodels.TimeOnly    `json:"start_time"`
+	EndTime         sharedmodels.TimeOnly    `json:"end_time"`
+	Duration        int                      `json:"duration"`
+	Days            shareddto.DaysFormView   `json:"days"`
+	EducatorIDs     string                   `json:"educator_ids"`
+	StudentIDs      string                   `json:"student_ids"`
+	URL             string                   `json:"url"`
+	Validation      map[string]events.Validation
+	StudentOptions  []sdto.StudentSelectBoxView
+	EducatorOptions []edto.EducatorSelectBoxView
 }
 
 func NewPeriodFormView(
 	p *models.Period,
 	allStudents []smodels.Student,
+	allEducators []emodels.Educator,
 ) PeriodFormView {
 	if p == nil {
 		return PeriodFormView{}
 	}
-	ssids := strings.Split(p.StudentIDs, ",")
 	return PeriodFormView{
-		ID:          p.ID,
-		Title:       p.Title,
-		ServiceType: p.ServiceType,
-		StartTime:   p.StartTime,
-		EndTime:     p.EndTime,
-		Duration:    p.Duration,
-		Days:        shareddto.DaysBitmaskToFormView(p.DaysBitmask),
-		EducatorIDs: p.EducatorIDs,
-		StudentIDs:  p.StudentIDs,
-		Validation:  events.Validate(p),
-		StudentList: sdto.NewStudentSelectBoxViews(allStudents, ssids),
+		ID:              p.ID,
+		Title:           p.Title,
+		ServiceType:     p.ServiceType,
+		StartTime:       p.StartTime,
+		EndTime:         p.EndTime,
+		Duration:        p.Duration,
+		Days:            shareddto.DaysBitmaskToFormView(p.DaysBitmask),
+		EducatorIDs:     strings.Join(p.EducatorIDs, ","),
+		StudentIDs:      strings.Join(p.StudentIDs, ","),
+		Validation:      events.Validate(p),
+		StudentOptions:  sdto.NewStudentSelectBoxViews(allStudents, p.StudentIDs),
+		EducatorOptions: edto.NewEducatorSelectBoxViews(allEducators, p.EducatorIDs),
 	}
 }
 
@@ -59,7 +61,7 @@ func (v PeriodFormView) ToPeriod() models.Period {
 		EndTime:     v.EndTime,
 		Duration:    v.Duration,
 		DaysBitmask: v.Days.ToBitmask(),
-		EducatorIDs: v.EducatorIDs,
-		StudentIDs:  v.StudentIDs,
+		EducatorIDs: strings.Split(v.EducatorIDs, ","),
+		StudentIDs:  strings.Split(v.StudentIDs, ","),
 	}
 }