+++ internal/features/students_periods/events/command_remove.go
@@ -76,7 +76,7 @@ func loadRemoveStudentFromPeriodContext(
 	*removeStudentFromPeriodContext,
 	error,
 ) {
-	query := streamQuery(periodID, studentID)
+	query := studentPeriodStreamQuery(periodID, studentID)
 	events, err := retriever.GetEvents(ctx, eventstore.NoEventPosition, 100, eventstore.Forward, query)
 	if err != nil {
 		return nil, err